Delete all previous solutions
This project was dead for a very long time and meanwhile I forgot all
details and concepts of ML and AI. So, to make up with what I had lost I
am beginning from scratch all over again, to watch all the tutorials
again and to do all the assignments again.

Thus, in the view of the above concern,
Delete all the previous solutions to start off virtually with a clean slate.

Signed-off-by: Suraj Upadhyay <usuraj35@gmail.com>